server/requirements.txt
fastapi
Pillow
uvicorn
deepface
python-multipart
python-jose[cryptography]
